Revert "Added logging function name on ServiceBus exception. (#162)"
This reverts commit cd7528ae4aab8ba90531a90e33e41b4e46cc4801.
