+++ release_notes.md
@@ -2,8 +2,9 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-#### Version Version 4.3.1
-- Added logging function name on ServiceBus exception.
+#### Version Version 4.3.0
+- Added support for allowing 'autoComplete' setting at the function level for all the single, batch or session triggers.
 
-**Release sprint:** Sprint 101
-[ [features](https://github.com/Azure/azure-functions-servicebus-extension/issues/162) ]
+
+**Release sprint:** Sprint 100
+[ [features](https://github.com/Azure/azure-functions-servicebus-extension/issues/138) ]
